backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
psycopg[binary]
yt-dlp>=2025.1.1
gunicorn==22.0.0
